[cleaner] Use compiled regex lists for parsers by default
This commit follows the initial change made to the username and keyword
parsers in #2823 and applies it to all parsers by default.

When a new match is found, a new `regex.Pattern()` object will now be
compiled and saved in the parser's map, and this object will be used for
ongoing obfuscations from that point forward, rather than rebuilding
regexes for every line we iterate over with that parser.

This is now built into the base `SoSMap` and leveraged by parsers,
rather than being handled directly by parsers. Further, this will be
enabled by default for all existing and new parsers. This shows decent
improvements to hostname parser performance in local testing. Note that
this functionality is explicitly disabled for the mac and ip parsers.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>