[tests] Minor fixes for avocado and unittests
`SoftwareManager()` as well as the actual `software_manager` module have
been reorganized within avocado, so we need to update our usage of them
accordingly.

Additionally, the legacy runner is being dismantled in avocado in favor
of the nrunner test runner. Upstream indicates it will be fully removed
sooner rather than later, so we should get ahead of it and switch to
nrunner. To do so successfully, since our tests are designed to have a
single `sos` execution per test case (Test() superclass), we need to
disable parallel execution of tests for this runner.

As part of local testing, it was found that the previous sysroot fix had
an unintentional side-effect of causing our unittests to fail if they
were run by a non-root user. This is because of a call to `os.chroot()`
in `_child_prep_fn()`. This has the edge case possibility of also
throwing an exception in certain invocations of `sos collect`, so add a
second conditional that we aren't trying to chroot to `/`. As a knock-on
effect, this will mean we are not constantly calling `os.chroot()` for
every command executed during report collections.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>