[report] Remove useless inclusion of tempfile_util in commons
An early implementation of #2729 included expanding `TempFileUtil` and
passing it as a means to create temp files to write command output
directly to. That approach was abandoned in favor of a more robust
implementation, but the cleanup of the previous design was apparently
not complete.

Fix this by removing the inclusion of tempfile_util in our `commons`
dict.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>