[runtimes] Allow container IDs to be used with `container_exists()`
As container runtimes can interchange container names and container IDs,
sos should also allow the use of container IDs when checking for the
presence of a given container.

In particular, this change unblocks the use of `Plugin.exec_cmd()` when
used in conjunction with `Plugin.get_container_by_name()` to pick a
container based on a provided regex that the container name may match.

Related: #2856

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>