[networking] list VRFs
Add command listing VRFs and their associated tables; including in
network namespaces. This information could be retrieved from reading the
`ip -d address show` output, but it is hard to see quickly. As the `ip
vrf show` output is quite small, let's include it.

If the system has two VRFs, the output will look like:

    Name              Table
    -----------------------
    vrf0                42
    vrf1                43

And if no VRF is defined the output will look like:

    Name              Table
    -----------------------
    No VRF has been configured

Signed-off-by: Antoine Tenart <atenart@kernel.org>