[tests] Add a test for non-default plugin_timeout compliance
There was a gap in our testing that allows #2863 to escape our notice -
that a `Plugin()`'s `plugin_timeout` attribute would be ignored if it
wasn't set to `TIMEOUT_DEFAULT`.

As that was resolved by #2864, add a test to ensure it remains working
as expected. The expected resolution order for a plugin's whole timeout
is as follows:

 1. The value set by `-k plugin.timeout`
 2. The value set by `--plugin-timeout`
 3. The value hardcoded in the plugin via the `plugin_timeout` attr
 4. `TIMEOUT_DEFAULT`

Related: #2863
Related: #2864

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>