[sosnode] Handle downstream versioning for runtime option check
First, adds parsing and formatting for an sos installation's release
version according to the loaded package manager for that node.

Adds a fallback version check for 4.2-13 for RHEL downstreams that
backport the `container-runtime` option into sos-4.2.

Carry this in upstream to account for use cases where a workstation used
to run `collect` from may be from a different stream than those used by
cluster nodes.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>