[sar] Limit sar conversion by age, remove shell encapsulation
The conversion of 'saX' binary data to human-readable 'sar' format is
typically handled by sysstat on the host system when a new day rolls
over the previous 'saX' file. As such, there should generally only be
one day (the current day's data) that needs manual conversion. For this
reason, limit our calling of 'sar' to files that have not been converted
already _and_ are less than a week old, to avoid spending time
converting very old files that are likely not useful.

Further, remove the shell encapsulation of the sar command.

Finally, add a docstring description of the plugin for `sos help`
output.

Closes: #1741

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>