[cleaner] Allow disabling specific parsers individually
Adds a new `--disable-parsers` option that allows users to selectively
disable parsers for a given execution of `sos clean`.

This may be useful in specific scenarios where obfuscation is not
strictly needed for all the types of data we obfuscate, and where the
user trusts whomever may be receiving the archive for review.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>