[docs] Correct automodule definitions for sphinx docs
Building sphinx docs was generating warnings for several locations about
sphinx being unable to reference certain classes.

This was found to be due to the `automodule` definitions for each
section including the `:noindex:` setting. Removing this configuration
resolved the vast majority of cross-referencing warnings that sphinx
would report on during a build of the docs.

Note that a single warning will remain, referencing threading.Thread,
after this change.

Closes: #2917

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>