[Plugin,Policy] Make a base tagging class, minor docstring fixes
Fixing the sphinx documentation warnings with the previous commit
unearthed a few inconsistencies with the docstrings for various bits of
`Plugin()` and `Policy()`. Fix those so that docs can be properly
formatted and cross-reference each other.

Related to this, add a new `PluginDistroTag` base class that tagging
classes will now subclass. At the moment, this only serves to enhance
the generate docs however this may prove useful in the future if we ever
want to leverage these taggings classes more directly.

Closes: #2917

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>