[report] --list-plugins should report used, not default, option values
When using `--list-plugins`, sos should report the values that will be
used in a given command, or with a given config file, not what the
default values are.

By reporting the set value, users can be sure their configuration or
commandline settings are being honored correctly before executing a
report collection.

Closes: #2921

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>