[GenericPolicy] Fix execution when GenericPolicy is loaded
Over time a lot has changed with `Policy()` implementation. Along with
that, the functionality of `GenericPolicy` broke at some point, thus
preventing any form of execution of SoS on distributions without
explicit policies.

Fix this by changing `GenericPolicy` to `GenericLinuxPolicy` (to be more
explicit in case we re-add non-Linux support in the future) and having
it subclass `LinuxPolicy` instead of the top-level `Policy`.

Further, add a platform check to ensure we are not inadvertently loading
a Linux-based policy on a non-Linux platform which would likely result
in failure to do anything useful.

The `GenericLinuxPolicy` will only load `IndependentPlugin`-tagged
plugins, does not leverage an init system, and does not provide any
default upload functionality.

Related: #2928

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>