[postgresql] Simplify val_type for dbport
After the change introduced via PR #2935
(Allow 'str' PlugOpt type to accept any value)
the postgresql plugin was the only one specifying
both str and int as possible values. This patch
simplifies the val_type field to be in line
with the rest of plugins where only one value
is used.

Signed-off-by: Jose Castillo <jcastillo@redhat.com>