[ceph_mon/ceph_osd] Collect ceph-mon backend type
Ceph uses leveldb (old versions) or rocksdb (current ones)
for metadata. The older filestore backend is deprecated since
Quincy (17.2.0) which uses leveldb. So it's useful to know
if, for example, ceph-mon is still using leveldb for metadata.
This can happen even after upgrading to newer Ceph versions
if the leveldb isn't migrated to leveldb.

Signed-off-by: Ponnuvel Palaniyappan <pponnuvel@gmail.com>