[host] Add collection of "top-level" symlinks and their targets
Adds collection of a `find` command that shows symlinks from /, or one
subdirectory below, that lists all symlinks and their targets in `ls`
output format. This is by request from a RH support team for situations
where problematic systems have directories underneathe / that are solely
symlinks, and this information is relevant to problem investigation.

While the `host` plugin may not be the most intuitive place for this
collection, there are no other existing plugins that would be "more
correct", and adding a new "symlinks" plugin would likely only serve to
confuse authors and maintainers going forward.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>