[clean] Don't re-create regex for parsers on every skip check
When determining if a parser should be skipped for a given file, we were
previously doing so ineffeciently by regenerating the regex object for
every file from the parser's `skip_files` class attr.

Instead, during parser initialization, use those string definitions to
create a list of regex objects to be used directly by the skip file
checks. This way, we only generate the regex objects once for each
parser for the entire run.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>