[logging] Safeguard against attempting to log non-UTF characters
The `logging` module does not support non UTF-8 characters, and as such
there was an exception being thrown when trying to log for example a
copied file that has such characters in the file name.

Safeguard against this by converting all messages logged via
`Plugin._format_msg()` to UTF-8, and replace any problematic characters
with `?`s to attempt to preserve context within the logs.

Closes: #2790

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>