[report] Add network device enumeration
Adds network device enumeration to the generated `devices` dict that
gets handed to plugins, and used for iteration by `add_device_cmd()`.

For the moment, this enumeration is limited to ethernet and bridge
devices, though this is expected to be expanded on later. These devices
are added to the top-level `network` element within the `devices` dict,
so both can be referenced at once using that key.

Additionally, namespaced network interfaces are added under a
`namespaced_network` key, which contains sub-dicts for each namespace
which in turn contain ethernet device lists. Note that these namespaced
device lists must be referenced specifically rather than via their key
strings, and this is done to avoid overlapping namespaced devices with
non-namespaced devices.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>