[redhat] Simplify `dist_version()` logic
Simplifies the logic in `dist_version()` to just return the major
version of the `redhat-release` package. This will not only remove a
requirement to update this check for new major version releases, but
also makes the call meaningful on RH-family distros aside from RHEL,
such as Fedora.

Note that this may fail on RHEL 5, but that is not a valid concern as
RHEL 5 does not support python3 and thus should never be in a situation
where sos-4.x+ is installed on such a system.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>