[teamd,tests] Add a test for teamd plugin and device enumeration
Adds a new test case for the `teamd` plugin that also inherently acts to
test team device enumeration.

Included with this test case is an addition to the base test classes
that allows for tests to define a `post_test_tear_down()` method that
will be run at the end of each test execution to allow for manual
cleanup - in this case deleting a 'fake' team device created for the
test.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>