Fix missing libvirt.all_logs option in libvirt plugin
Seems like there's use of 'all_logs' option in the code of the libvirt
plugin, but there's no definition of it as an available option in the
class. This change just adds it so that an external user can use it.

I tested it with:
```
sudo `which sos` report -o libvirt  -k libvirt.all_logs
```

and by checking:

```
sudo `which sos` report -l | grep "libvirt.all_logs
```

Signed-off-by: Osher De Paz <odepaz@redhat.com>