[snapcraft] Update base to core22
core20 has python3-magic 0.4.15, and hence the binary detection will
not work by default. Moving to core22 goes to 0.4.24-2 which would
mean this would work nevertheless. This also means migrating to using
craftctl instead of snapcraftctl.

Snap build now needs to be done on 22.04, as it core22, so chaging
the snap build for cirrus to the corresponding image. The version
cannot be set now if adopt-info is not there, so removing this for
the cirrus tests.

Signed-off-by: Arif Ali <arif.ali@canonical.com>