[testing] Fine tune the testing for ubuntu
There was no testing for ubuntu 18.04, and previous manual testing found
issues, and this is still supported. Add envs similar to fedora and centos
testing. This change also ensures that ubuntu 22.04 is the one that is used
for snap builds. We don't need python3-magic on the OS, as this should now
come in-built within the snap, which will have the right version for binary
checks.

Update images to the latest ones available for Ubuntu

Signed-off-by: Arif Ali <arif.ali@canonical.com>