[firewall_tables] Add enablement triggers
Adds enablement triggers for the `firewall_tables` plugin in the form of
the `files` and `kernel_mods` tuples. This means that the plugin will no
longer blindly run by default, and only run if either the
`/etc/nftables/` directory is present or if the relevant kernel modules
are loaded. The latter should help in our efforts to ensure that we do
not inadvertantly load new kernel modules during execution.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>