[krb5] Enable on Ubuntu/Debian, expand package tuple
It was found that the `krb5` plugin was not executing on Ubuntu/Debian
systems due to a legacy change that inadvertently removed support for
those distributions.

Re-enable support by defining a new plugin class for those distros,
setting the kdc directory appropriately. Additionally, expand the
package tuple to include newer package names.

Closes: #3041

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>