[report,Plugin] Add a tag summary to report manifest
After all plugins have run, sos will now generate a "tag summary" and
add it to the report section of the manifest. This summary will be a
json-formatted entry that details all collections within the report that
have any tag associated with them at all. This should allow for easier
parsing of the manifest by analyzers such as Red Hat Insights.

As part of this change, commands will no longer be automatically tagged
with the name of the binary used in the command collection.

Additionally, manual collections performed by a plugin's `collect()`
method will now be recorded in the manifest in the same manner as file
and command output collections.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>