[testing] Switch to CodeQL from LGTM
LGTM is shutting down shortly, as it has been folded into GH CodeQL. As
such, we need to switch over to using CodeQL to maintain automatic code
analysis.

Do this by using the provided CodeQL workflow configuration, and add a
config file to limit the analysis to the appropriate `sos/` directory.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>