[report] Merge redundant conditional in calling `display_results()`
During our final processing to report the results of an archive to the
user at the end of an execution, there was a redundant conditional
checking if we were running with `--build` to decide how to call
`display_results()` (with or without stat information). Merge this into
the previous conditional that determines that stat information, which in
turn resolves a CodeQL error alert.

Related: #3066

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>