[networking] Nest namespace iteration inside namespace conditional
A CodeQL error alert was highlighted that the iptables predicates might
be unset during execution of the plugin, which in turn could potentially
cause the plugin to throw an exception. Fix this by nesting the
interation of command collections over namespaces that might use this
predicate in the same conditional that defines the predicate.

Related: #3066

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>