[openstack_mistral] Collect mistral content itself
Logs as well as deploy information are located in /var/lib/mistral as
well. Until now, this location was overlooked, and important information
about the deploy state, configuration and potential errors were missing
from the initial SOS-Report, leading to time loss for both Customer and
Support.

This patch intends to correct this situation, by ensuring the content is
taken from the Undercloud.

Notes:
- this is especially important for OSP<17.0
- the location may content multiple subdirectories
- the location will contain the history of the different actions done by
  the operator
- mistral logs are moved from the openstack_instack plugin to the new
  openstack_mistral for the sake of consistency. The "instack" name is
  deprecated in OSP.

Signed-off-by: Cédric Jeanneret <cjeanner@redhat.com>