[ceph] Add new collect clust profile for RHCS5
Adds a new cluster profile for RHCS 5 for `sos collect`. This profile
depends upon the use of `cephadm` which is used to both deploy and
manage the cluster. Users may optionally restrict the list of nodes to
collect from by using the `-c ceph.labels` option to specify a set of
label(s) to filter node results with.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>