[packit] Update packit configuration for Copr srpm builds
Packit is moving srpm builds from sandbox to copr, which means we need
to specify the build dependencies for packit to know how to create the
build environment.

As part of this change, re-organize the config file to be easier to
read, and update some of the deprecated keys or nesting per docs.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>