[ceph_mon] Update plugin for newer ceph versions
Newer versions of ceph do not have running processes directly on the
node. As such, the current method of determining the ID for the `ceph
tell` commands is not functional. Fix this by using the output of `ceph
status` on RHCS 5 nodes to get the node IDs.

Further, while ceph containers can be used to enable this plugin, those
containers by default cannot run various `ceph` commands collected by
the plugin. However, those commands are functional directly on the host,
so no longer attempt to execute within the containers.

Finally, update `postproc()` for newer versions as well.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>