[ceph_osd] Update plugin for newer versions of ceph
Newer versions of Ceph, such as for RHCS 5, make some changes to the
container names as well as how to leverage `ceph daemon` commands.
Update the plugin to reflect this, and use the available admin sockets
on the host to capture that output, rather than always running it inside
the osd container.

Further, account for the use of an fsid in directory paths for newer ceph.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>